yes, this is vba
Yes, this is ugly, yes, the code is tightly bound to a
particular problem, but it's one of those instances
where you just need to get a job done.

As to why I'm putting it out there — I'd like to have a
collection of copy-pasteable snippets lying around for future use.
With simple automation like this, time is often limited, and
the faster I churn out something at least close to workable,
the more valuable it becomes.